+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Serve favicon as embedded resource. [#468](https://github.com/elastic/package-registry/pull/468)
 * Generate index.json file. [#470](https://github.com/elastic/package-registry/pull/470)
 * Stream archived package content. [#472](https://github.com/elastic/package-registry/pull/472)
+* Generate package index.json files. [#479](https://github.com/elastic/package-registry/pull/479)
 
 ### Deprecated